db_service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
python-dotenv==1.0.0
requests==2.31.0
pydantic==2.4.2
numpy==1.26.4 